style: Add copyrights to all project files

+++ SDialogs/src/main/java/smith/lib/alerts/dialog/AlertSDialog.java
@@ -1,3 +1,21 @@
+/*
+ *
+ *   Created by Dev. Smith (Hussein Shakir) on 9/16/23, 10:07 PM
+ *   Copyright Ⓒ 2023. All rights reserved Ⓒ 2023 http://github.com/smith8h
+ *   Last modified: 8/16/23, 12:38 AM
+ *
+ *   Licensed under the Apache License, Version 2.0 (the "License");
+ *   you may not use this file except in compliance with the License.
+ *                     You may obtain a copy of the License at
+ *                  http://www.apache.org/licenses/LICENSE-2.0
+ *   Unless required by applicable law or agreed to in writing,
+ *   software distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ * /
+ */
+
 package smith.lib.alerts.dialog;
 
 import android.annotation.SuppressLint;
@@ -20,7 +38,7 @@ public class AlertSDialog extends SDialog {
     private int maxHeight;
 
     /**
-     * Pass the current context you using this sdialog from.
+     * Pass the current context you're using this sdialog from.
      * @param context Current context (or Activity).
      */
     @SuppressLint("InflateParams")
@@ -137,10 +155,10 @@ public class AlertSDialog extends SDialog {
 
     /**
      * Set a max height to fix screen overriding content.
-     * @param maxHeight an int value as maxHeight (it will automatically converted to dp);
+     * @param maxHeight an int value as maxHeight (it will automatically convert to dp);
      */
     public void setMaxHeight(int maxHeight) {
-        this.maxHeight = (int) utils.dp(maxHeight);
+        this.maxHeight = utils.dp(maxHeight);
     }
 
     /**
